requirements.txt
flask==2.1.1
flask-sqlalchemy
pymysql
WTForms~=2.3.3
cryptography
